+++ sickbeard/show_queue.py
@@ -897,7 +897,7 @@ class QueueItemRefresh(ShowQueueItem):
         logger.log('Performing refresh on %s' % self.show_obj.name)
 
         self.show_obj.refresh_dir()
-        self.show_obj.write_metadata()
+        self.show_obj.write_metadata(force=self.force)
         # if self.force:
         #    self.show_obj.update_metadata()
         self.show_obj.populate_cache(self.force_image_cache)